% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyprojmarks-package.R
\docType{package}
\name{psyprojmarks-package}
\alias{psyprojmarks}
\alias{psyprojmarks-package}
\title{psyprojmarks: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
